[fix] 修复mknod/map/磁盘设置太小/无法自动创建文件/linux目录项问题

See merge request T202510486995194/os2025!32

+++ include/kernel/fs/file.h
@@ -70,7 +70,7 @@ struct file_vnode {
 struct file {
     enum { FD_NONE, FD_PIPE, FD_REG, FD_DEVICE } f_type;
     uint8 f_mode;         ///< 访问模式
-    uint8 f_flags;        ///< 打开文件时的标志（如O_APPEND等）
+    uint f_flags;         ///< 打开文件时的标志（如O_APPEND等）
     uint64 f_pos;         ///< 偏移量
     uint16 f_count;       ///< 引用计数，表示有多少用户或进程持有此文件结构
     short f_major;        ///< 设备号（如果是设备文件）
